Move create squad tests to service tests
